- changed x-axis values in the tour chart from float to double, this was necessary to show the history chart with better accuracy when the virtual chart width is 1'000'000'000 pixel
- implementing day/second units in the history x-axis

+++ net.tourbook.chart/src/net/tourbook/chart/Chart.java
@@ -810,7 +810,7 @@ public class Chart extends ViewForm {
 				.getValuesIndex());
 	}
 
-	public int getXXDevViewPortLeftBorder() {
+	public long getXXDevViewPortLeftBorder() {
 		return _chartComponents.getChartComponentGraph().getXXDevViewPortLeftBorder();
 	}
